Added interface
